Unify the func_0805FA60 prototype

+++ include/mode_sio2.h
@@ -196,7 +196,7 @@ void func_0801227C(void);
 void func_080125A4(void);
 void ColliderSetRadius(void* a, u16 b);
 void UpdatePlayTime(void);
-void func_0805FA60(s32 a, void* b, s32 c, s32 d);
+void func_0805FA60(s32 a, void* b, s32 c, u8 d);
 void func_0805FA8C(s32 a, u16 b, u16 c);
 void func_0805FC04(u8 x, u8 y, u8 c, u16 v);
 void func_0805FCB0(u8 x, u8 y, u8 c, const char* s);
